requirements.txt
PyQt5==5.15.7
transformers==4.30.2
pillow==9.5.0
torch==1.13.1
